requirements.txt
python-telegram-bot==20.6
fastapi==0.111.0
uvicorn==0.30.1
pydub==0.25.1
deep-translator==1.11.4
